+++ src/main/java/org/hisp/dhis/integration/t2a/model/AnalyticsGrid.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.dhis2.integration.model;
+package org.hisp.dhis.integration.t2a.model;
 
 import java.util.List;
 
@@ -35,27 +35,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @Data
 @JsonIgnoreProperties( ignoreUnknown = true )
-public class ProgramIndicatorGroup
+public class AnalyticsGrid
 {
-    private List<ProgramIndicator> programIndicators;
-
-    public String programIndicatorAsString()
-    {
-        StringBuilder result = new StringBuilder();
-
-        int x = 0;
-
-        for ( ProgramIndicator pi : programIndicators )
-        {
-            result.append( pi.getId() ).append( ";" );
-
-            // todo remove break
-            if ( x++ == 3 )
-            {
-                break;
-            }
-        }
-
-        return result.deleteCharAt( result.length() - 1 ).toString();
-    }
+    private List<List<String>> rows;
 }
