+++ src/main/java/org/hisp/dhis/integration/t2a/processors/PostDataValueSetQueryBuilder.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.dhis2.integration.processors;
+package org.hisp.dhis.integration.t2a.processors;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -33,8 +33,8 @@ import java.util.Set;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
-import org.dhis2.integration.model.*;
-import org.dhis2.integration.routes.T2ARouter;
+import org.hisp.dhis.integration.t2a.model.*;
+import org.hisp.dhis.integration.t2a.routes.T2ARouter;
 
 public class PostDataValueSetQueryBuilder implements Processor
 {
