+++ src/main/java/org/hisp/dhis/integration/t2a/routes/T2ARouterNew.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.dhis2.integration.routes;
+package org.hisp.dhis.integration.t2a.routes;
 
 import java.util.Base64;
 import java.util.Optional;
@@ -35,8 +35,9 @@ import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.builder.ThreadPoolBuilder;
 import org.apache.camel.component.properties.PropertiesComponent;
-import org.dhis2.integration.model.ProgramIndicatorGroup;
-import org.dhis2.integration.processors.*;
+import org.hisp.dhis.integration.t2a.model.ProgramIndicatorGroup;
+import org.hisp.dhis.integration.t2a.processors.AnalyticsGridQueryBuilder;
+import org.hisp.dhis.integration.t2a.processors.AnalyticsGridToDataValueSetQueryBuilder;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
